[INT]Segregated classes to run as different nodes
