+++ src/main.cpp
@@ -60,19 +60,5 @@ int main(int argc, char** argv) {
   else
     ROS_INFO("The robot could not reach the first goal for some reason");
 }
-/*
-#include <visualization_msgs/Marker.h>
-#include "nav_msgs/Odometry.h"
-#include "location.h"
 
-int main(int argc, char *argv[]) {
-  ros::init(argc,argv,"vmarkers");
-  ros::NodeHandle nh;
-   
-  Location vmarker;
-  vmarker.visualizeLocation(argc, argv);
 
-  return 0;
-}
-
-*/
